Clean up composables
- delete useAmountBasedDirectDebitDependentFormAction.ts and useAmountBasedFormAction.ts
- move useFormAction in DE Desktop to src/components/composables and name it useFormActionWithReceipt.ts
- add tests for the `useFormActionWithReceipt.ts` composable

- also extract the `useFormAction` component in the EN banner and replace it with the shared `useFormActionWithReceipt`

- add documentation

https://phabricator.wikimedia.org/T383980
