+++ banners/english/C25_WMDE_Desktop_EN_00/components/BannerVar.vue
@@ -116,7 +116,7 @@ import DoubleProgressBar from '@src/components/ProgressBar/DoubleProgressBar.vue
 import SoftClose from '@src/components/SoftClose/SoftClose.vue';
 import WMDEFundsForwardingEN from '@src/components/UseOfFunds/Infographics/WMDEFundsForwardingEN.vue';
 import { useBannerHider } from '@src/components/composables/useBannerHider';
-import { useFormAction } from '../useFormAction';
+import { useFormActionWithReceipt } from '@src/components/composables/useFormActionWithReceipt';
 import { FormActions } from '@src/domain/FormActions';
 
 const minimumAmount = 10;
@@ -149,7 +149,7 @@ const stepControllers = [
 	createSubmittableUpgradeToYearly( formModel, FormStepNames.MainDonationFormStep, FormStepNames.MainDonationFormStep )
 ];
 
-const { formAction } = useFormAction( inject<FormActions>( 'formActions' ), minimumAmount );
+const { formAction } = useFormActionWithReceipt( inject<FormActions>( 'formActions' ), minimumAmount );
 
 watch( contentState, async () => {
 	emit( 'bannerContentChanged' );
